% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{clr_samples_rows}
\alias{clr_samples_rows}
\title{Calculate the centered log transform (clr) of an otu table.}
\usage{
clr_samples_rows(x, constant = 0.5, log = 2)
}
\description{
This function takes an OTU counts table as input returns the clr transform.
Samples must be rows, and columns as otus. The default log base is 2.
Log base can also be set to 10 or natural log
}
\examples{
## example
otu_tab1 <- matrix(
c(2, 4, 3, 1, 5, 7),
nrow=2, ncol=3)
rownames(otu_tab1) <- c("sample1","sample2")
colnames(otu_tab1) <- c("otu1","otu2","otu3")

clr_samples_rows(otu_tab1)
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data,
Monographs on Statistics and Applied Probability. Chapman & Hall Ltd.,
London (UK). 416p.
}

