% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{prop_table}
\alias{prop_table}
\title{Convert OTU table to a proportion table}
\usage{
prop_table(x, Margin = 1)
}
\description{
This function takes an OTU counts table as input returns a proportion table.
The default input is samples as rows (Margin=1), but samples can be set to
columns by setting Margin to 2.
}
\examples{
## example otu table with samples as rows
otu_tab1 <- matrix(
c(2, 4, 3, 1, 5, 7),
nrow=2, ncol=3)
rownames(otu_tab1) <- c("sample1","sample2")
colnames(otu_tab1) <- c("otu1","otu2","otu3")

prop_table(otu_tab1)
}

