% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{rarefy_table}
\alias{rarefy_table}
\title{Rarefy OTU table to a specificied sample depth}
\usage{
rarefy_table(x, depth = 1000, Margin = 1)
}
\description{
This function takes an OTU counts table as input returns a subsampled table.
The default input is samples as rows (Margin=1), but samples can be set to
columns by setting Margin to 2. The rarified table is a multinomial sample
with probabilities set to observed counts witihn each sample. The depth is
set by user, often is the min observed depth (i.e. min(rowSums(otu_table))).
}
\examples{
## example otu table with samples as rows
otu_tab1 <- matrix(
c(2, 4, 3, 1, 5, 7),
nrow=2, ncol=3)
rownames(otu_tab1) <- c("sample1","sample2")
colnames(otu_tab1) <- c("otu1","otu2","otu3")

rarefy_table(otu_tab1)
}

